requirements.txt
groq
python-dotenv
pydantic
tenacity